Runnable workflows UI (#503)
* lockfile and gitignore

* workflow data structure

* deps

* Refactor into separate files

* Extract getOperationDetails from WorkflowDetail

* Rough initial implementation

* Tweaks to the @fiberplane/hono lib

* Wire up the output logic

* Cleanup

* Ui cleanup

* please the linting overlords

* Add handling "number"

---------
